build: Move paths into group to SVG transform
A transform is required to convert from the glyph coordinate
system to the SVG "upside-down" coordinate system. It was
applied on the <svg> tag itself, but that is not valid according
to the SVG spec. Now all objects under the root <svg> tag are
grouped in a <g> to apply the transform.

Fixes #4